my_project/HOMEWORK4/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv
redis
